Fix: renames printers option to formatters and does resolution in bin
